feat(encoding): write out intermediate points on the SGD
Ideally this should be behind a feature flag.

TODO: serialize the numpy array as something other than a JSON array of 18 arrays of 512 floats.
This ASCII output is not the most optimal encoding (~200k each, versus 36k for 9k float32s), but serialization time is minimal compared to the 170ms - 300ms execution time on a GPU to perform the SGD.